Added a simple-and-always-allowed cast to provide class compatibility even to java 1.5
git-svn-id: file:///home/toad/git-migration/temprepository/trunk/freenet@14883 67a373e5-eb02-0410-a15c-ee090a768436

+++ src/freenet/support/TimeSortedHashtable.java
@@ -32,7 +32,7 @@ public class TimeSortedHashtable {
 	private class MyComparator implements Comparator {
 
 		public int compare(Object arg0, Object arg1) {
-			if(arg0 instanceof Long && arg1 instanceof Long) return ((Long)arg0).compareTo(arg1);
+			if(arg0 instanceof Long && arg1 instanceof Long) return ((Long)arg0).compareTo((Long)arg1);
 			if(arg0 instanceof Element && arg1 instanceof Element) return ((Element)arg0).compareTo(arg1);
 			// Comparing a Long with an Element, because we are searching for an Element by the value of a Long.
 			// Hence we do not need to consider the element value.
